fix(leetcode): save non-playground iframe
previously we just skip those and render nothing in the document, which works fine for videos in solution articles because of the context.

I realized there are other random iframes too (like DB Fiddles in SQL articles). My goal in general is to scrape all info offline if possible, but for obscure cases like this I don't think it's worth it to write custom scraper for it. Therefore this commit will save the link and attempt to display the iframe in the document, just in case. (videos won't work tho)